split up base files, move into new directory
