+++ src/parsers/citi-costco-credit-card-parser.ts
@@ -1,11 +1,8 @@
 import {parsePageItems} from 'pdf-text-reader';
 import {DEBUG} from '../config';
-import {
-    createParserStateMachine,
-    ParsedOutput,
-    ParsedTransaction,
-    StatementParser,
-} from '../parser-base/base-parser';
+import {ParsedOutput, ParsedTransaction} from '../parser-base/parsed-output';
+import {createParserStateMachine} from '../parser-base/parser-state-machine';
+import {StatementParser} from '../parser-base/statement-parser';
 import {getPdfDocument} from '../readPdf';
 import {flatten2dArray} from '../util/array';
 import {dateFromSlashFormat, dateWithinRange} from '../util/date';
